requirements/base.txt
dj-database-url==0.4.2
Django==1.11.2
django-rest-swagger==2.1.2
djangorestframework==3.6.3
psycopg2==2.7.1
requests==2.18.1
